admin func but movie del

+++ back-end/routes/api_v1.php
@@ -80,9 +80,9 @@ Route::middleware('auth:sanctum')->group(function () {
   
   //* Only Admin func
   Route::middleware('restrictToAdmin')->group(function() {
-    Route::delete('/movies/{movieId}', [MovieController::class, 'deleteMovie']);
-    Route::put('/add-moderator', [AdminController::class, 'addModerator']);
-    Route::put('/delete-moderator', [AdminController::class, 'deleteModerator']);
+    Route::delete('/movies/admin/{movieId}', [MovieController::class, 'deleteMovie']);
+    Route::put('/admin/add-moderator', [AdminController::class, 'addModerator']);
+    Route::put('/admin/delete-moderator', [AdminController::class, 'deleteModerator']);
     Route::put('/admin/ban', [AdminController::class, 'banUser']);
     Route::put('/admin/unban', [AdminController::class, 'unbanUser']);
   });
